fix: パスの調整
